+++ src/main/java/domain/repository/INTERFACES/TaxiLiveAddressRepository.java
@@ -7,5 +7,5 @@ import domain.repository.DAO.INTERFACES.IBaseDao;
 import java.util.List;
 
 public interface TaxiLiveAddressRepository extends IBaseDao<TaxiLiveAddress> {
-    List<Cab> findNearbyCabs(String pointWKT, int meters_distance);
+    List<Cab> findNearbyCabs(String pointWKT, double meters_distance);
 }
